@@ -7,26 +7,31 @@ const fs = require('fs');
 const path = require('path');
 
 patch(`./dist/apps/electron/node_modules/@angular/cli/lib/cli/index.js`);
+patch(`./dist/apps/vscode/node_modules/@angular/cli/lib/cli/index.js`);
 
-fs.readdirSync('./dist/packages').forEach(dir => {
-  try {
-    patch(
-      path.join(
-        './dist/packages',
-        dir,
-        'resources',
-        'app',
-        'node_modules/@angular/cli/lib/cli/index.js'
-      )
-    );
-  } catch (e) {}
-});
+try {
+  fs.readdirSync('./dist/packages').forEach(dir => {
+    try {
+      patch(
+        path.join(
+          './dist/packages',
+          dir,
+          'resources',
+          'app',
+          'node_modules/@angular/cli/lib/cli/index.js'
+        )
+      );
+    } catch (e) {}
+  });
+} catch (e) {}
 
 function patch(path) {
-  const file = fs.readFileSync(path).toString();
-  const patched = file.replace(
-    "const [, localPath] = config_1.getWorkspaceRaw('local');",
-    'const localPath = null;'
-  );
-  fs.writeFileSync(path, patched);
+  try {
+    const file = fs.readFileSync(path).toString();
+    const patched = file.replace(
+      "const [, localPath] = config_1.getWorkspaceRaw('local');",
+      'const localPath = null;'
+    );
+    fs.writeFileSync(path, patched);
+  } catch (e) {}
 }
